feat: init command
