--- packages/plugin-trpc/cli/src/commands/trpc/hello-world.ts
@@ -1,16 +1,0 @@
-import { GluegunCommand } from '@webstone/gluegun';
-
-const command: GluegunCommand = {
-	name: 'hello',
-	alias: ['h'],
-	description: 'Hello World Command',
-	hidden: false,
-	dashed: false,
-	run: async (toolbox) => {
-		const { print } = toolbox;
-
-		print.info(`Hello World`);
-	}
-};
-
-export default command;
