+++ packages/plugin-trpc/cli/src/templates/base/router.ts.ejs
@@ -1,0 +1,7 @@
+import { router, publicProcedure } from './trpc';
+
+const appRouter = router({
+	greeting: publicProcedure.query(() => 'hello tRPC v10!')
+});
+
+export type AppRouter = typeof appRouter;
