+++ packages/plugin-trpc/cli/src/templates/base/trpc.ts.ejs
@@ -1,0 +1,7 @@
+import { initTRPC } from 'webstone-plugin-web-trpc';
+const t = initTRPC.create();
+
+export const middleware = t.middleware;
+export const router = t.router;
+export const publicProcedure = t.procedure;
+export const mergeRouters = t.mergeRouters;
